requirement.txt
asgiref==3.3.1
Django==3.1.7
django-fav==0.3
django-filter==2.4.0
django-phone-field==1.8.1
djangorestframework==3.12.4
Markdown==3.3.4
Pillow==8.1.2
pytz==2021.1
sqlparse==0.4.1
